# Conflicts:
#	Packages/manifest.json
#	Packages/packages-lock.json

+++ Assets/Script/Structs/GameResultData.cs
@@ -5,9 +5,9 @@ public struct GameResultData
 {
     public int Day;
     public int MaxCombo;
-    public int Score;
+    public float Score;
      
-    public GameResultData(int day, int maxCombo, int score)
+    public GameResultData(int day, int maxCombo, float score)
     {
         Day = day;
         MaxCombo = maxCombo;
